Add gh-ost support
