+++ src/ServiceProvider.php
@@ -2,6 +2,7 @@
 
 namespace Daursu\ZeroDowntimeMigration;
 
+use Daursu\ZeroDowntimeMigration\Connections\GhostConnection;
 use Daursu\ZeroDowntimeMigration\Connections\PtOnlineSchemaChangeConnection;
 use Illuminate\Database\Connection;
 use Illuminate\Database\Connectors\MySqlConnector;
@@ -30,7 +31,15 @@ class ServiceProvider extends IlluminateServiceProvider
             return new PtOnlineSchemaChangeConnection($connection, $database, $prefix, $config);
         });
 
+        Connection::resolverFor('gh-ost', function ($connection, $database, $prefix, $config) {
+            return new GhostConnection($connection, $database, $prefix, $config);
+        });
+
         $this->app->bind('db.connector.pt-online-schema-change', function () {
+            return new MySqlConnector;
+        });
+
+        $this->app->bind('db.connector.gh-ost', function () {
             return new MySqlConnector;
         });
     }
